Hack for renamed constructors
SWIG doesn't namespace "renamed" constructors with the class's symbolic name
(it relied on the constructor name being the class symbolic name). This is an issue
for all languages (try %rename on a constructor whose class is instantiated multiple times,
or that shares a name with a free function, or with another class's constructor).

The commented-out patch would probably fix this in general, but it breaks python's
shadow wrapping because "sym:name" is reset before the python wrappers are made.